webApp - code refactor
