+++ justbuild.it-WebApp/src/main/java/justbuild/it/web/app/entity/enums/ServiceCategory.java
@@ -1,4 +1,4 @@
-package justbuild.it.web.app.model;
+package justbuild.it.web.app.entity.enums;
 
 public enum ServiceCategory {
