+++ justbuild.it-WebApp/src/main/java/justbuild/it/web/app/service/OfferCreationService.java
@@ -1,6 +1,6 @@
 package justbuild.it.web.app.service;
 
-import justbuild.it.web.app.model.Offer;
+import justbuild.it.web.app.entity.Offer;
 import justbuild.it.web.app.repository.OfferFileRepository;
 import org.springframework.stereotype.Service;
